+++ CHANGELOG.md
@@ -30,9 +30,13 @@ for further information about branching and tagging conventions.
 
 #### Added
 - Error message when input amount is less than min amount
+- New build configuration
 
 #### Changed
 - REFERENCE_MAX_LENGTH from 255 to 64
+
+#### Removed
+- `SupportedBrowsers` component
 
 ## [1.11.0-rc.0] - 2020-02-05
 #### Fixed